+++ CHANGELOG.md
@@ -62,7 +62,8 @@
 * feat(PageGrid): new component
 * feat(PageColumns): new component
 * feat(PageList): new component
-* feat(components): expose `ui` in slot props where used
+* feat(components): expose `b24ui` in slot props where used
+* feat(module): add `experimental.componentDetection` option
 * fix(ProseLi,ProseOl,ProseUl,ProseP): fix color && size
 * fix(ProseH1\H2\H3\H4\H5\H6): add support hash
 * fix(SidebarLayout): improve theme, import Button, add slot 'content-right'